GH-264: Integer division instead of double in GetSpectralResolution.

+++ Sources/Accord.Audio/Tools.cs
@@ -155,7 +155,7 @@ namespace Accord.Audio
         /// 
         public static double GetSpectralResolution(int samplingRate, int samples)
         {
-            return samplingRate / samples;
+            return samplingRate / (double)samples;
         }
 
         /// <summary>
